% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrsreport-package.R
\docType{package}
\name{scrsreport-package}
\alias{scrsreport}
\alias{scrsreport-package}
\title{SCRS Statistical Report Bookdown Project Template}
\description{
This package adds a project template to RStudio. It uses bookdown,
together with current SCRS LaTeX and Markdown templates to facilitate a
professtional looking statistical report.
}
\author{
\strong{Maintainer}: Tan Tran \email{tan.tran@montana.edu} (0000-0001-9881-6339)

Authors:
\itemize{
  \item Kenneth Flagg (0000-0003-3136-4766)
  \item Jordan Schupbach
  \item Michael Lerch
}

Other contributors:
\itemize{
  \item  MSU SCRS [copyright holder, funder]
}

}
\keyword{internal}
